软件串口完活了

+++ MDK-ARM/SoftWareUART/SoftWareUart.c
@@ -63,13 +63,6 @@ SoftWareUart_Status_t SoftWareUart_Sendbyte(SoftWareUart_Handle_t handle, uint8_
     return UART_OK;
 }
 
-/**
- * @brief
- *
- * @param handle
- * @param buffer
- * @param size
- */
 SoftWareUart_Status_t SoftWareUart_SendBuffer(SoftWareUart_Handle_t handle, const uint8_t *buffer, size_t size)
 {
     if (handle == NULL)
@@ -140,15 +133,15 @@ void SoftWareUart_TimeCallback(SoftWareUart_Handle_t *handle)
     else
     {
         // 接收结束了（通常第9次是停止位，可以检查也可以直接结束）
-        if(uart->rxbuffer != NULL)
+        if (uart->rxbuffer != NULL)
         {
             uart->rxbuffer[uart->rx_write_index++] = uart->recvData;
             if (uart->rx_write_index >= uart->rx_size)
             {
                 uart->rx_write_index = 0; // 循环
             }
-			
-			(*handle)->flag.Recivedata = 1;
+
+            (*handle)->flag.Recivedata = 1;
         }
         HAL_TIM_Base_Stop_IT(uart->HardWare.InterruptHtime); // 停止计时器中断
         uart->recvStat = COM_STOP_BIT;                       // 重置状态，准备下一次接收
@@ -166,14 +159,16 @@ void SoftWareUart_RXCallback(SoftWareUart_Handle_t *handle)
     {
         if (uart->recvStat == COM_STOP_BIT) // 空闲状态，发现起始位
         {
-            uart->recvStat = COM_START_BIT;                       // 标记开始采样
-            uart->recvData = 0;                                   // 清空上次接收的数据
-            if(!uart->flag.Firstdata)
+            uart->recvStat = COM_START_BIT; // 标记开始采样
+            uart->recvData = 0;             // 清空上次接收的数据
+            if (!uart->flag.Firstdata)
             {
                 uart->flag.Firstdata = 1;
-                SoftWareUart_delayus(uart,uart->baud * 1.5); 
-            } else {
-                SoftWareUart_delayus(uart,uart->baud / 2);        					 // 跳过起始位等待
+                SoftWareUart_delayus(uart, uart->baud * 1.5);
+            }
+            else
+            {
+                SoftWareUart_delayus(uart, (int)(uart->baud / 2)); // 跳过起始位等待
             }
 
             HAL_TIM_Base_Start_IT(uart->HardWare.InterruptHtime); // 开启定时器中断，定时采样
@@ -183,22 +178,24 @@ void SoftWareUart_RXCallback(SoftWareUart_Handle_t *handle)
 
 SoftWareUart_Status_t SoftWareUart_CheckRevice(SoftWareUart_Handle_t handle)
 {
-	if(handle == NULL) return UART_ERROR;
-	
-	if(handle->flag.Recivedata)
-	{
-	return UART_OK;
-	}
-	return UART_ERROR;
+    if (handle == NULL)
+        return UART_ERROR;
+
+    if (handle->flag.Recivedata)
+    {
+        return UART_OK;
+    }
+    return UART_ERROR;
 }
 
 SoftWareUart_Status_t SoftWareUart_Clearbuffer(SoftWareUart_Handle_t *handle)
 {
 
-if(handle == NULL || *handle == NULL) return UART_ERROR;
-	(*handle)->flag.Recivedata = 0;
-	memset((*handle)->rxbuffer,0,(*handle)->rx_size);
-	(*handle)->recvData = 0x00;
-	(*handle)->rx_write_index = 0;
-	return UART_OK;
+    if (handle == NULL || *handle == NULL)
+        return UART_ERROR;
+    (*handle)->flag.Recivedata = 0;
+    memset((*handle)->rxbuffer, 0, (*handle)->rx_size);
+    (*handle)->recvData = 0x00;
+    (*handle)->rx_write_index = 0;
+    return UART_OK;
 }
